Update README.md
A just a small edit on the getting started steps. I went ahead and blindly created an index in pinecone not knowing what I should set the dimensions to, so I arbitrarily put in a number, which later came back to bite me because it wasn't in line with OpenAI's 1536. I only found this out after it failed ingestion. If I had known to start with 1536 from the start, I could have avoided this. 

I know you mention this as a debug step in the bottom, but placing it right at the start could avoid this issue in the first place.